Homework-2 modified
A kért változtatások implementálása.
- terminal_cleaner.py
- többsoros print
- list slicing alkalmazása
