+++ homeworks/editerdelyhorvath/2_vars_datatypes/terminal_clearer.py
@@ -1,0 +1,7 @@
+import os
+
+def clear_terminal():
+    if os.name == 'nt':  # Windows
+        os.system('cls')
+    else:  # Mac and Linux
+        os.system('clear')
